修改:技术交易:删除改为状态删除;逻辑修改;BUG修复

+++ technology-transaction-service/src/main/java/com/itts/technologytransactionservice/service/ITJsShService.java
@@ -19,9 +19,9 @@ public interface ITJsShService extends IService<TJsSh> {
 
 	IPage page(Query query);
 
-	TJsSh selectBycgxqId(Long cgxqId);
+	TJsSh selectBycgxqId(Integer cgxqId);
 
-	List<TJsSh> selectBycgxqIds(List<Long> cgxqIds);
+	List<TJsSh> selectBycgxqIds(List<Integer> cgxqIds);
 
 	/*int updateById(Long id);*/
 }
