+++ technology-transaction-service/src/main/java/com/itts/technologytransactionservice/service/impl/TJsShServiceImpl.java
@@ -28,12 +28,12 @@ public class TJsShServiceImpl extends ServiceImpl<TJsShMapper, TJsSh> implements
 	}
 
 	@Override
-	public TJsSh selectBycgxqId(Long cgxqId) {
+	public TJsSh selectBycgxqId(Integer cgxqId) {
 		return tJsShMapper.selectBycgxqId(cgxqId);
 	}
 
-	public List<TJsSh> selectBycgxqIds(List<Long> cgxqIds) {
-		Long[] objects = cgxqIds.toArray(new Long[cgxqIds.size()]);
+	public List<TJsSh> selectBycgxqIds(List<Integer> cgxqIds) {
+		Integer[] objects = cgxqIds.toArray(new Integer[cgxqIds.size()]);
 		return tJsShMapper.selectBycgxqIds(objects);
 	}
